AcostaAPI/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
colorama==0.4.6
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
pydantic==2.6.1
pydantic_core==2.16.2
PyMySQL==1.1.0
pyodbc==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1